// src/api/stetsom/model/cmsProductDetailPayload.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Stetsom API
 * API documentation for Stetsom
 * OpenAPI spec version: 1.0.0
 */
import type { Category } from "./category";
import type { Product } from "./product";

export interface CmsProductDetailPayload {
  product: Product;
  category: Category;
}