// src/api/stetsom/model/localeInput.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Stetsom API
 * API documentation for Stetsom
 * OpenAPI spec version: 1.0.0
 */

export type LocaleInput = (typeof LocaleInput)[keyof typeof LocaleInput];

export const LocaleInput = {
  pt: "pt",
  en: "en",
  es: "es",
} as const;