// src/api/stetsom/model/productStatus.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Stetsom API
 * API documentation for Stetsom
 * OpenAPI spec version: 1.0.0
 */

export type ProductStatus = (typeof ProductStatus)[keyof typeof ProductStatus];

export const ProductStatus = {
  PUBLISHED: "PUBLISHED",
  DRAFT: "DRAFT",
  SCHEDULED: "SCHEDULED",
} as const;